--- test/agents/.gitignore
@@ -1,2 +1,0 @@
-venv/
-.env
