--- test/agents/guest.py
@@ -1,35 +1,0 @@
-from dotenv import load_dotenv
-from livekit.agents import (
-    Agent,
-    AgentSession,
-    JobContext,
-    WorkerOptions,
-    cli,
-)
-from livekit.plugins import deepgram, elevenlabs, openai, silero
-
-
-load_dotenv(dotenv_path=".env", override=True)
-
-
-async def entrypoint(ctx: JobContext):
-    await ctx.connect()
-
-    agent = Agent(
-        instructions="You are a guest on a podcast."
-                     " The audio from this conversation will be streamed to live listeners."
-                     " Choose a field of study and expertise, and talk about recent developments in that field.",
-        turn_detection="vad",
-    )
-    session = AgentSession(
-        vad=silero.VAD.load(),
-        stt=deepgram.STT(model="nova-3"),
-        llm=openai.LLM(model="gpt-4o-mini"),
-        tts=elevenlabs.TTS(),
-    )
-
-    await session.start(agent=agent, room=ctx.room)
-
-
-if __name__ == "__main__":
-    cli.run_app(WorkerOptions(entrypoint_fnc=entrypoint, agent_name="egress-integration-guest"))
