--- test/agents/host.py
@@ -1,36 +1,0 @@
-from dotenv import load_dotenv
-from livekit.agents import (
-    Agent,
-    AgentSession,
-    JobContext,
-    WorkerOptions,
-    cli,
-)
-from livekit.plugins import deepgram, elevenlabs, openai, silero
-
-
-load_dotenv(dotenv_path=".env", override=True)
-
-
-async def entrypoint(ctx: JobContext):
-    await ctx.connect()
-
-    agent = Agent(
-        instructions="You are hosting a podcast, and you will be having a conversation with your guest."
-                     " The audio from this conversation will be streamed to live listeners."
-                     " Ask engaging questions and keep the conversation flowing.",
-        turn_detection="vad"
-    )
-    session = AgentSession(
-        vad=silero.VAD.load(),
-        stt=deepgram.STT(model="nova-3"),
-        llm=openai.LLM(model="gpt-4o-mini"),
-        tts=elevenlabs.TTS(),
-    )
-
-    await session.start(agent=agent, room=ctx.room)
-    await session.generate_reply(instructions="Greet your guest and ask them about their field of study.")
-
-
-if __name__ == "__main__":
-    cli.run_app(WorkerOptions(entrypoint_fnc=entrypoint, agent_name="egress-integration-host"))
