+++ test/edge.go
@@ -20,7 +20,6 @@ import (
 	"context"
 	"fmt"
 	"math/rand"
-	"os"
 	"testing"
 	"time"
 
@@ -56,17 +55,6 @@ func (r *Runner) testEdgeCases(t *testing.T) {
 					fileType: livekit.EncodedFileType_OGG,
 				},
 				custom: r.testRoomCompositeLateTrackDuration,
-			},
-
-			// Agents with room composite audio only
-
-			{
-				name:        "Agents",
-				requestType: types.RequestTypeRoomComposite,
-				fileOptions: &fileOptions{
-					filename: "agents_{time}",
-				},
-				custom: r.testAgents,
 			},
 
 			// RoomComposite audio mixing
@@ -232,7 +220,7 @@ func (r *Runner) testRoomCompositeLateTrackDuration(t *testing.T, test *testCase
 	}, lksdk.NewRoomCallback())
 	require.NoError(t, err)
 	t.Cleanup(p2.Disconnect)
-	publishLegacyTrack(t, p2.LocalParticipant, types.MimeTypeOpus, make(chan struct{}))
+	PublishOnRoom(t, p2, 1, true, false)
 
 	// Let the late track record for a few seconds
 	time.Sleep(time.Second * 7)
@@ -258,17 +246,6 @@ func (r *Runner) testRoomCompositeLateTrackDuration(t *testing.T, test *testCase
 		"file duration should not exceed wall-clock duration (inflated by late track offset)")
 }
 
-func (r *Runner) testAgents(t *testing.T, test *testCase) {
-	_, err := os.Stat("/agents/.env")
-	if err != nil {
-		t.Skip("skipping agents test; missing env file")
-	}
-
-	r.launchAgents(t)
-	time.Sleep(time.Second * 5)
-	r.runFileTest(t, test)
-}
-
 func (r *Runner) testAudioMixing(t *testing.T, test *testCase) {
 	p1, err := lksdk.ConnectToRoom(r.WsUrl, lksdk.ConnectInfo{
 		APIKey:              r.ApiKey,
@@ -279,7 +256,7 @@ func (r *Runner) testAudioMixing(t *testing.T, test *testCase) {
 	}, lksdk.NewRoomCallback())
 	require.NoError(t, err)
 	t.Cleanup(p1.Disconnect)
-	publishLegacyTrack(t, p1.LocalParticipant, types.MimeTypeOpus, make(chan struct{}))
+	PublishOnRoom(t, p1, 0, true, false)
 
 	agent, err := lksdk.ConnectToRoom(r.WsUrl, lksdk.ConnectInfo{
 		APIKey:              r.ApiKey,
@@ -291,7 +268,7 @@ func (r *Runner) testAudioMixing(t *testing.T, test *testCase) {
 	}, lksdk.NewRoomCallback())
 	require.NoError(t, err)
 	t.Cleanup(agent.Disconnect)
-	publishLegacyTrack(t, agent.LocalParticipant, types.MimeTypeOpus, make(chan struct{}))
+	PublishOnRoom(t, agent, 2, true, false)
 
 	p2, err := lksdk.ConnectToRoom(r.WsUrl, lksdk.ConnectInfo{
 		APIKey:              r.ApiKey,
@@ -302,7 +279,7 @@ func (r *Runner) testAudioMixing(t *testing.T, test *testCase) {
 	}, lksdk.NewRoomCallback())
 	require.NoError(t, err)
 	t.Cleanup(p2.Disconnect)
-	publishLegacyTrack(t, p2.LocalParticipant, types.MimeTypeOpus, make(chan struct{}))
+	PublishOnRoom(t, p2, 1, true, false)
 
 	r.runFileTest(t, test)
 }
@@ -351,8 +328,7 @@ func (r *Runner) testRoomCompositeStaysOpen(t *testing.T, test *testCase) {
 	require.NoError(t, err)
 	r.room = room
 
-	publishLegacyTrack(t, r.room.LocalParticipant, types.MimeTypeOpus, make(chan struct{}))
-	publishLegacyTrack(t, r.room.LocalParticipant, types.MimeTypeVP8, make(chan struct{}))
+	PublishOnRoom(t, r.room, 0, true, true)
 
 	time.Sleep(time.Second * 10)
 
@@ -535,11 +511,11 @@ func (r *Runner) testSrtFailure(t *testing.T, test *testCase) {
 }
 
 func (r *Runner) testTrackDisconnection(t *testing.T, test *testCase) {
-	pub := publishLegacyTrack(t, r.room.LocalParticipant, types.MimeTypeVP8, make(chan struct{}))
-	test.videoTrackID = pub.SID()
+	_, videoPub := PublishOnRoom(t, r.room, 0, false, true)
+	test.videoTrackID = videoPub.SID()
 
 	time.AfterFunc(time.Second*10, func() {
-		pub.SimulateDisconnection(time.Second * 10)
+		videoPub.SimulateDisconnection(time.Second * 10)
 	})
 
 	r.runFileTest(t, test)
