+++ test/sync_analyzer.go
@@ -20,11 +20,14 @@ import (
 	"bufio"
 	"bytes"
 	"context"
+	"encoding/csv"
 	"errors"
 	"fmt"
+	"io"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"regexp"
 	"strconv"
 	"strings"
 	"testing"
@@ -34,6 +37,476 @@ import (
 
 	"github.com/livekit/protocol/logger"
 )
+
+//----------------------------------------------------------------------
+// Shared constants, regex patterns, and utility functions
+//----------------------------------------------------------------------
+
+const (
+	testSampleSilenceLevel = -38
+	testSampleBeepLevel    = -30.0
+)
+
+var (
+	rePTS  = regexp.MustCompile(`pts_time:([0-9.]+)`)
+	reYAVG = regexp.MustCompile(`lavfi\.signalstats\.YAVG[=:]\s*([0-9.]+)`)
+	reRMS  = regexp.MustCompile(`RMS_level[=:](-?[0-9.infINFNaN]+)`)
+)
+
+func parsePTSSecondsToDuration(s string) (time.Duration, error) {
+	f, err := strconv.ParseFloat(strings.TrimSpace(s), 64)
+	if err != nil {
+		return 0, err
+	}
+	return time.Duration(f * float64(time.Second)), nil
+}
+
+func secondsToDuration(f float64) time.Duration {
+	return time.Duration(f * float64(time.Second))
+}
+
+func averageSpacing(ts []time.Duration) (time.Duration, error) {
+	if len(ts) < 2 {
+		return 0, fmt.Errorf("need at least 2 timestamps (got %d)", len(ts))
+	}
+
+	var sum time.Duration
+	var gaps int
+	for i := 1; i < len(ts); i++ {
+		d := ts[i] - ts[i-1]
+		if d <= 0 {
+			continue
+		}
+		sum += d
+		gaps++
+	}
+	if gaps == 0 {
+		return 0, fmt.Errorf("no positive gaps to compute spacing")
+	}
+	return time.Duration(int64(sum) / int64(gaps)), nil
+}
+
+func requireDurationInDelta(t *testing.T, expected, actual, delta time.Duration, msgAndArgs ...interface{}) {
+	require.InDelta(t,
+		expected.Nanoseconds(),
+		actual.Nanoseconds(),
+		float64(delta.Nanoseconds()),
+		msgAndArgs...)
+}
+
+func absDuration(d time.Duration) time.Duration {
+	if d < 0 {
+		return -d
+	}
+	return d
+}
+
+//----------------------------------------------------------------------
+// FFmpeg helpers — run ffmpeg/ffprobe and extract stats to log files
+//----------------------------------------------------------------------
+
+func ffmpegVideoStats(videoPath, statsFile string) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
+	defer cancel()
+
+	cmd := exec.CommandContext(ctx, "ffmpeg",
+		"-hide_banner", "-nostats", "-loglevel", "repeat+info",
+		"-i", videoPath,
+		"-map", "0:v:0",
+		"-vf", fmt.Sprintf("crop=w=iw:h=8:x=0:y=0,signalstats,metadata=print:file=%s", statsFile),
+		"-f", "null", "-")
+
+	var outBuf, errBuf bytes.Buffer
+	cmd.Stdout = &outBuf
+	cmd.Stderr = &errBuf
+
+	if err := cmd.Run(); err != nil {
+		if errors.Is(ctx.Err(), context.DeadlineExceeded) {
+			return fmt.Errorf("ffmpeg video stats timeout after 15s")
+		}
+		return fmt.Errorf("ffmpeg video stats extraction failed: %w\nstdout:\n%s\nstderr:\n%s",
+			err, outBuf.String(), errBuf.String())
+	}
+	return nil
+}
+
+func ffmpegAudioStats(audioPath, statsFile string) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
+	defer cancel()
+
+	cmd := exec.CommandContext(ctx, "ffmpeg",
+		"-hide_banner", "-nostats", "-loglevel", "repeat+info",
+		"-i", audioPath,
+		"-af", fmt.Sprintf("pan=mono|c0=0.5*c0+0.5*c1,astats=metadata=1:reset=1,ametadata=print:key=lavfi.astats.Overall.RMS_level:file=%s", statsFile),
+		"-f", "null", "-")
+
+	var outBuf, errBuf bytes.Buffer
+	cmd.Stdout = &outBuf
+	cmd.Stderr = &errBuf
+
+	if err := cmd.Run(); err != nil {
+		if errors.Is(ctx.Err(), context.DeadlineExceeded) {
+			return fmt.Errorf("ffmpeg audio stats timeout after 15s")
+		}
+		return fmt.Errorf("ffmpeg audio stats extraction failed: %w\nstdout:\n%s\nstderr:\n%s",
+			err, outBuf.String(), errBuf.String())
+	}
+	return nil
+}
+
+func ffmpegSilenceStats(audioPath string, noiseLevel int, minDuration float64) (*bytes.Buffer, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
+	defer cancel()
+
+	cmd := exec.CommandContext(ctx, "ffmpeg",
+		"-hide_banner", "-nostats", "-loglevel", "info",
+		"-i", audioPath,
+		"-af", "silencedetect=noise="+fmt.Sprintf("%d", noiseLevel)+"dB:d="+strconv.FormatFloat(minDuration, 'f', -1, 64),
+		"-f", "null", "-")
+
+	var stderr bytes.Buffer
+	cmd.Stderr = &stderr
+
+	if err := cmd.Run(); err != nil {
+		if errors.Is(ctx.Err(), context.DeadlineExceeded) {
+			return nil, fmt.Errorf("ffmpeg silence stats timeout after 15s")
+		}
+		return nil, fmt.Errorf("ffmpeg silence stats extraction failed: %w\nstderr:\n%s",
+			err, stderr.String())
+	}
+	return &stderr, nil
+}
+
+//----------------------------------------------------------------------
+// Single-participant content extraction (legacy test media)
+//----------------------------------------------------------------------
+
+// extractFlashTimestamps runs ffmpeg + signalstats on the top stripe
+// and returns one timestamp per flash event (YAVG >= 130, spaced >= 0.2s).
+func extractFlashTimestamps(videoPath, outPath string) ([]time.Duration, error) {
+	logFile := filepath.Join(outPath, "video_flash.log")
+
+	err := ffmpegVideoStats(videoPath, logFile)
+	if err != nil {
+		return nil, err
+	}
+
+	file, err := os.Open(logFile)
+	if err != nil {
+		return nil, fmt.Errorf("ffmpeg video stats failed to open log file: %w", err)
+	}
+	defer file.Close()
+
+	const flashThreshold = 130.0
+	const minGap = 200 * time.Millisecond
+
+	var (
+		flashes   []time.Duration
+		lastFlash = -999 * time.Second
+		curPTS    time.Duration
+	)
+
+	scanner := bufio.NewScanner(file)
+	for scanner.Scan() {
+		line := scanner.Text()
+
+		if m := rePTS.FindStringSubmatch(line); len(m) == 2 {
+			if d, perr := parsePTSSecondsToDuration(m[1]); perr == nil {
+				curPTS = d
+			}
+			continue
+		}
+
+		if m := reYAVG.FindStringSubmatch(line); len(m) == 2 {
+			y, _ := strconv.ParseFloat(m[1], 64)
+			if y >= flashThreshold && curPTS-lastFlash > minGap {
+				flashes = append(flashes, curPTS)
+				lastFlash = curPTS
+			}
+		}
+	}
+	return flashes, scanner.Err()
+}
+
+// extractBeepTimestamps runs ffmpeg + astats to find beeps.
+// A beep is when RMS_level > beepThreshold, debounced by 0.2s.
+func extractBeepTimestamps(audioPath string, beepThreshold float64, outPath string) ([]time.Duration, error) {
+	logFile := filepath.Join(outPath, "audio_beep.log")
+
+	err := ffmpegAudioStats(audioPath, logFile)
+	if err != nil {
+		return nil, err
+	}
+
+	file, err := os.Open(logFile)
+	if err != nil {
+		return nil, fmt.Errorf("ffmpeg audio stats failed to open log file: %w", err)
+	}
+	defer file.Close()
+
+	const minGap = 200 * time.Millisecond
+
+	var (
+		beeps  []time.Duration
+		last   = -999 * time.Second
+		curPTS time.Duration
+	)
+
+	scanner := bufio.NewScanner(file)
+	for scanner.Scan() {
+		line := scanner.Text()
+
+		if m := rePTS.FindStringSubmatch(line); len(m) == 2 {
+			if d, perr := parsePTSSecondsToDuration(m[1]); perr == nil {
+				curPTS = d
+			}
+			continue
+		}
+
+		if m := reRMS.FindStringSubmatch(line); len(m) == 2 {
+			val := m[1]
+			if strings.Contains(val, "inf") || strings.Contains(val, "nan") {
+				continue
+			}
+			lvl, _ := strconv.ParseFloat(val, 64)
+			if lvl > beepThreshold && curPTS-last > minGap {
+				beeps = append(beeps, curPTS)
+				last = curPTS
+			}
+		}
+	}
+	return beeps, scanner.Err()
+}
+
+// silenceRange represents one silence segment.
+type silenceRange struct {
+	start    time.Duration
+	end      time.Duration
+	duration time.Duration
+}
+
+// detectSilence runs ffmpeg silencedetect and returns all silence ranges.
+func detectSilence(audioPath string, noiseLevel int, minDuration time.Duration) ([]silenceRange, error) {
+	stderr, err := ffmpegSilenceStats(audioPath, noiseLevel, minDuration.Seconds())
+	if err != nil {
+		return nil, err
+	}
+
+	var ranges []silenceRange
+	var current silenceRange
+	inSilence := false
+
+	reStart := regexp.MustCompile(`silence_start:\s*([0-9.]+)`)
+	reEnd := regexp.MustCompile(`silence_end:\s*([0-9.]+)\s*\|\s*silence_duration:\s*([0-9.]+)`)
+
+	scanner := bufio.NewScanner(stderr)
+	for scanner.Scan() {
+		line := scanner.Text()
+
+		if m := reStart.FindStringSubmatch(line); len(m) == 2 {
+			if start, perr := strconv.ParseFloat(m[1], 64); perr == nil {
+				current = silenceRange{start: secondsToDuration(start)}
+				inSilence = true
+			}
+		}
+
+		if m := reEnd.FindStringSubmatch(line); len(m) == 3 {
+			if inSilence {
+				end, _ := strconv.ParseFloat(m[1], 64)
+				dur, _ := strconv.ParseFloat(m[2], 64)
+				current.end = secondsToDuration(end)
+				current.duration = secondsToDuration(dur)
+				ranges = append(ranges, current)
+				inSilence = false
+			}
+		}
+	}
+
+	return ranges, scanner.Err()
+}
+
+//----------------------------------------------------------------------
+// Content verification — replaces the old content_checks.go functions.
+// Called from verifyFile / verifySegments / verifyStream.
+//----------------------------------------------------------------------
+
+// verifyContent runs standard content checks on an output file.
+// For audio+video: extracts flashes and beeps, checks they exist.
+// For video-only: checks flash count matches duration and spacing is ~1s.
+// For audio-only: extracts beeps, checks they exist.
+func verifyContent(t *testing.T, file string, info *FFProbeInfo, hasAudio, hasVideo bool, muting bool, outPath string) {
+	t.Helper()
+
+	if muting {
+		// content checks not yet reliable with muting enabled
+		return
+	}
+
+	if hasVideo {
+		flashes, err := extractFlashTimestamps(file, outPath)
+		require.NoError(t, err)
+		logger.Debugw("flashes", "flashes", flashes, "count", len(flashes))
+
+		if !hasAudio && info != nil {
+			// video-only: verify flash count and spacing
+			dur, err := parseFFProbeDuration(info.Format.Duration)
+			require.NoError(t, err)
+			require.InDelta(t, len(flashes), dur.Round(time.Second).Seconds(), 3)
+			avgFlashSpacing, err := averageSpacing(flashes)
+			require.NoError(t, err)
+			requireDurationInDelta(t, avgFlashSpacing, time.Second, time.Millisecond*200)
+		}
+	}
+
+	if hasAudio {
+		beeps, err := extractBeepTimestamps(file, testSampleBeepLevel, outPath)
+		require.NoError(t, err)
+		logger.Debugw("beeps", "beeps", beeps, "count", len(beeps))
+
+		silenceRanges, err := detectSilence(file, testSampleSilenceLevel, time.Millisecond*100)
+		if len(silenceRanges) > 0 || err != nil {
+			logger.Errorw("silence ranges not empty", err, "silenceRanges", silenceRanges)
+		}
+	}
+}
+
+//----------------------------------------------------------------------
+// Special-case content checks (used as testCase.contentCheck)
+//----------------------------------------------------------------------
+
+// verifyVideoGap checks that flashes have a ~10s gap starting around t=10s
+// (video unpublished at 10s, republished at 20s).
+func verifyVideoGap(t *testing.T, file string, info *FFProbeInfo) {
+	flashes, err := extractFlashTimestamps(file, filepath.Dir(file))
+	require.NoError(t, err)
+
+	dur, err := parseFFProbeDuration(info.Format.Duration)
+	require.NoError(t, err)
+
+	gapLength := time.Second * 10
+	require.InDelta(
+		t,
+		float64(len(flashes))+gapLength.Seconds(),
+		dur.Round(time.Second).Seconds(),
+		5.0,
+		"flashes+gap ~= duration (±5s)",
+	)
+
+	gapsFound := 0
+	for i := 1; i < len(flashes); i++ {
+		if flashes[i]-flashes[i-1] > gapLength-time.Millisecond*500 {
+			gapsFound++
+			requireDurationInDelta(t, flashes[i], time.Second*20, time.Second*2)
+		} else {
+			requireDurationInDelta(t, flashes[i], flashes[i-1], time.Second+time.Millisecond*200)
+		}
+	}
+	require.Equal(t, gapsFound, 1)
+}
+
+//----------------------------------------------------------------------
+// Keyframe interval verification (for stream tests)
+//----------------------------------------------------------------------
+
+func keyframeContentCheck(expectedInterval float64) func(t *testing.T, target string, _ *FFProbeInfo) {
+	return func(t *testing.T, target string, _ *FFProbeInfo) {
+		requireKeyframeInterval(t, target, expectedInterval)
+	}
+}
+
+func requireKeyframeInterval(t *testing.T, input string, expectedInterval float64) {
+	t.Helper()
+	if expectedInterval <= 0 {
+		return
+	}
+
+	timestamps, err := ffprobeKeyframeTimestamps(input, expectedInterval)
+
+	require.NoError(t, err)
+	require.GreaterOrEqual(t, len(timestamps), 2, "ffprobe returned less than two keyframes for %s", input)
+
+	tolerance := 0.020 // 20ms
+	prev := timestamps[0]
+	found := false
+	for _, ts := range timestamps[1:] {
+		if ts <= prev {
+			prev = ts
+			continue
+		}
+		found = true
+		require.InDelta(t, expectedInterval, ts-prev, tolerance, "keyframe spacing mismatch for %s", input)
+		prev = ts
+	}
+	require.True(t, found, "no increasing keyframe timestamps found for %s", input)
+}
+
+func ffprobeKeyframeTimestamps(input string, expectedInterval float64) ([]float64, error) {
+	timestamps := []float64{}
+	var err error
+
+	readSeconds := expectedInterval*4 + 1
+
+	args := []string{
+		"-v", "error",
+		"-fflags", "nobuffer",
+		"-rw_timeout", "5000000",
+		"-select_streams", "v:0",
+		"-show_packets",
+		"-show_entries", "packet=pts_time,dts_time,flags,stream_index,size,pos",
+		"-of", "csv=p=0",
+		input,
+	}
+
+	timeout := time.Duration(readSeconds) * time.Second
+	ctx, cancel := context.WithTimeout(context.Background(), timeout)
+	defer cancel()
+
+	cmd := exec.CommandContext(ctx, "ffprobe", args...)
+	stdout, err := cmd.StdoutPipe()
+	if err != nil {
+		return nil, fmt.Errorf("stdout pipe: %w", err)
+	}
+
+	if err = cmd.Start(); err != nil {
+		return nil, fmt.Errorf("start ffprobe: %w", err)
+	}
+	defer cmd.Wait()
+
+	csvReader := csv.NewReader(stdout)
+
+	for {
+		record, e := csvReader.Read()
+		if e != nil {
+			if ctx.Err() == nil && e != io.EOF {
+				err = fmt.Errorf("read csv: %w", e)
+			}
+			break
+		}
+
+		if len(record) != 6 {
+			err = fmt.Errorf("unexpected record length: %d", len(record))
+			break
+		}
+
+		pts, e := strconv.ParseFloat(record[1], 64)
+		if e != nil {
+			err = fmt.Errorf("parse pts: %w", e)
+			break
+		}
+		if strings.Contains(record[5], "K") {
+			timestamps = append(timestamps, pts)
+		}
+	}
+
+	if err != nil {
+		return nil, err
+	}
+	return timestamps, nil
+}
+
+//----------------------------------------------------------------------
+// Multi-participant sync analysis
+//----------------------------------------------------------------------
 
 // ParticipantSpec describes a participant for sync analysis.
 type ParticipantSpec struct {
@@ -66,15 +539,13 @@ type videoEvent struct {
 }
 
 // extractFrequencyAudio runs FFmpeg with a bandpass filter at the given frequency
-// and returns timestamps where RMS exceeds the threshold.
-// This isolates a single participant's audio from the composited mix.
+// and returns time ranges where RMS exceeds the threshold.
 func extractFrequencyAudio(audioPath string, frequency float64, outPath string) ([]audioEvent, error) {
 	logFile := filepath.Join(outPath, fmt.Sprintf("audio_%dhz.log", int(frequency)))
 
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 	defer cancel()
 
-	// Bandpass filter isolates the target frequency, then astats extracts RMS levels
 	af := fmt.Sprintf("bandpass=f=%.0f:width_type=h:w=100,astats=metadata=1:reset=1,ametadata=print:key=lavfi.astats.Overall.RMS_level:file=%s",
 		frequency, logFile)
 
@@ -100,7 +571,7 @@ func extractFrequencyAudio(audioPath string, frequency float64, outPath string) 
 }
 
 // parseAudioEvents reads an astats log file and groups consecutive above-threshold
-// frames into events. Returns a list of (start, end) duration pairs.
+// frames into events.
 func parseAudioEvents(logFile string, threshold float64) ([]audioEvent, error) {
 	file, err := os.Open(logFile)
 	if err != nil {
@@ -130,7 +601,6 @@ func parseAudioEvents(logFile string, threshold float64) ([]audioEvent, error) {
 		if m := reRMS.FindStringSubmatch(line); len(m) == 2 {
 			val := m[1]
 			if strings.Contains(val, "inf") || strings.Contains(val, "nan") {
-				// silence — close any open event
 				if current != nil {
 					current.end = curPTS
 					events = append(events, *current)
@@ -145,7 +615,6 @@ func parseAudioEvents(logFile string, threshold float64) ([]audioEvent, error) {
 				}
 				current.end = curPTS
 			} else if current != nil {
-				// Below threshold — if gap is small, keep the event open (debounce)
 				if curPTS-current.end > minGap {
 					events = append(events, *current)
 					current = nil
@@ -161,15 +630,13 @@ func parseAudioEvents(logFile string, threshold float64) ([]audioEvent, error) {
 }
 
 // extractRegionFlashes runs FFmpeg with a crop filter targeting a specific tile region
-// and detects flash events using signalstats YAVG, matching extractFlashTimestamps() logic.
+// and detects flash events using signalstats YAVG.
 func extractRegionFlashes(videoPath string, tileX, tileY, tileW, tileH int, outPath string) ([]videoEvent, error) {
 	logFile := filepath.Join(outPath, fmt.Sprintf("video_region_%d_%d.log", tileX, tileY))
 
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 	defer cancel()
 
-	// Crop to the tile region's top 8 pixels (matching existing flash detection),
-	// then run signalstats to get YAVG
 	vf := fmt.Sprintf("crop=w=%d:h=8:x=%d:y=%d,signalstats,metadata=print:file=%s",
 		tileW, tileX, tileY, logFile)
 
@@ -196,7 +663,6 @@ func extractRegionFlashes(videoPath string, tileX, tileY, tileW, tileH int, outP
 }
 
 // parseFlashEvents reads a signalstats log file and returns timestamps of flash events.
-// Uses the same YAVG >= 130 threshold and 200ms minimum gap as extractFlashTimestamps().
 func parseFlashEvents(logFile string) ([]videoEvent, error) {
 	file, err := os.Open(logFile)
 	if err != nil {
@@ -236,11 +702,7 @@ func parseFlashEvents(logFile string) ([]videoEvent, error) {
 }
 
 // AnalyzeSync verifies intra-participant and inter-participant A/V sync
-// in a composited output file. It fails the test if any participant's offset
-// exceeds the tolerance.
-//
-// The rotation schedule is assumed: participant i is active (audio unmuted)
-// from t = i*turnDuration to t = (i+1)*turnDuration, cycling through all participants.
+// in a composited output file.
 func AnalyzeSync(
 	t *testing.T,
 	outputFile string,
@@ -253,7 +715,6 @@ func AnalyzeSync(
 
 	results := make([]SyncResult, len(participants))
 
-	// Extract per-frequency audio events and per-region video flashes
 	allAudioEvents := make([][]audioEvent, len(participants))
 	allVideoEvents := make([][]videoEvent, len(participants))
 
@@ -268,16 +729,12 @@ func AnalyzeSync(
 		results[i].Participant = p.Identity
 	}
 
-	// Analyze each participant's turn
 	n := len(participants)
 	for i, p := range participants {
 		windowStart := time.Duration(i) * turnDuration
 		windowEnd := windowStart + turnDuration
 
-		// --- Intra-participant sync ---
-		// Find the first audio event that overlaps this participant's active window
 		audioOnset := findAudioOnset(allAudioEvents[i], windowStart, windowEnd)
-		// Find the first video flash in this window
 		videoOnset := findVideoOnset(allVideoEvents[i], windowStart, windowEnd)
 
 		results[i].AudioPresent = audioOnset >= 0
@@ -300,8 +757,6 @@ func AnalyzeSync(
 				p.Identity, audioOnset, videoOnset, intraOffset)
 		}
 
-		// --- Inter-participant sync ---
-		// Compare actual audio onset to expected schedule
 		if audioOnset >= 0 {
 			expectedOnset := windowStart
 			interOffset := absDuration(audioOnset - expectedOnset)
@@ -315,7 +770,6 @@ func AnalyzeSync(
 				"tolerance", tolerance,
 			)
 
-			// Inter-participant tolerance is more relaxed — mute propagation adds latency
 			interTolerance := tolerance + 500*time.Millisecond
 			require.LessOrEqual(t, interOffset, interTolerance,
 				"inter-participant sync exceeded tolerance for %s: expected=%v actual=%v offset=%v",
@@ -327,14 +781,12 @@ func AnalyzeSync(
 			)
 		}
 
-		_ = n // used by rotation wrap-around logic if needed
+		_ = n
 	}
 
 	return results
 }
 
-// findAudioOnset returns the start time of the first audio event that overlaps
-// the given window, or -1 if none found.
 func findAudioOnset(events []audioEvent, windowStart, windowEnd time.Duration) time.Duration {
 	for _, e := range events {
 		if e.end >= windowStart && e.start <= windowEnd {
@@ -347,8 +799,6 @@ func findAudioOnset(events []audioEvent, windowStart, windowEnd time.Duration) t
 	return -1
 }
 
-// findVideoOnset returns the timestamp of the first video flash within the window,
-// or -1 if none found.
 func findVideoOnset(events []videoEvent, windowStart, windowEnd time.Duration) time.Duration {
 	for _, e := range events {
 		if e.timestamp >= windowStart && e.timestamp <= windowEnd {
@@ -356,11 +806,4 @@ func findVideoOnset(events []videoEvent, windowStart, windowEnd time.Duration) t
 		}
 	}
 	return -1
-}
-
-func absDuration(d time.Duration) time.Duration {
-	if d < 0 {
-		return -d
-	}
-	return d
 }
